metrics/influxdb: reuse code between v1 and v2 reporters (#26963)
