+++ metrics/influxdb/influxdbv2.go
@@ -2,7 +2,6 @@ package influxdb
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	"github.com/XinFinOrg/XDPoSChain/log"
@@ -78,145 +77,13 @@ func (r *v2Reporter) run() {
 func (r *v2Reporter) send() {
 	r.reg.Each(func(name string, i interface{}) {
 		now := time.Now()
-		namespace := r.namespace
-
-		switch metric := i.(type) {
-
-		case metrics.Counter:
-			v := metric.Count()
-
-			measurement := fmt.Sprintf("%s%s.count", namespace, name)
-			fields := map[string]interface{}{
-				"value": v,
-			}
-
-			pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-			r.write.WritePoint(pt)
-
-		case metrics.CounterFloat64:
-			v := metric.Count()
-
-			measurement := fmt.Sprintf("%s%s.count", namespace, name)
-			fields := map[string]interface{}{
-				"value": v,
-			}
-
-			pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-			r.write.WritePoint(pt)
-
-		case metrics.Gauge:
-			ms := metric.Snapshot()
-
-			measurement := fmt.Sprintf("%s%s.gauge", namespace, name)
-			fields := map[string]interface{}{
-				"value": ms.Value(),
-			}
-
-			pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-			r.write.WritePoint(pt)
-
-		case metrics.GaugeFloat64:
-			ms := metric.Snapshot()
-
-			measurement := fmt.Sprintf("%s%s.gauge", namespace, name)
-			fields := map[string]interface{}{
-				"value": ms.Value(),
-			}
-
-			pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-			r.write.WritePoint(pt)
-
-		case metrics.Histogram:
-			ms := metric.Snapshot()
-
-			if ms.Count() > 0 {
-				ps := ms.Percentiles([]float64{0.5, 0.75, 0.95, 0.99, 0.999, 0.9999})
-				measurement := fmt.Sprintf("%s%s.histogram", namespace, name)
-				fields := map[string]interface{}{
-					"count":    ms.Count(),
-					"max":      ms.Max(),
-					"mean":     ms.Mean(),
-					"min":      ms.Min(),
-					"stddev":   ms.StdDev(),
-					"variance": ms.Variance(),
-					"p50":      ps[0],
-					"p75":      ps[1],
-					"p95":      ps[2],
-					"p99":      ps[3],
-					"p999":     ps[4],
-					"p9999":    ps[5],
-				}
-
-				pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-				r.write.WritePoint(pt)
-			}
-
-		case metrics.Meter:
-			ms := metric.Snapshot()
-
-			measurement := fmt.Sprintf("%s%s.meter", namespace, name)
-			fields := map[string]interface{}{
-				"count": ms.Count(),
-				"m1":    ms.Rate1(),
-				"m5":    ms.Rate5(),
-				"m15":   ms.Rate15(),
-				"mean":  ms.RateMean(),
-			}
-
-			pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-			r.write.WritePoint(pt)
-
-		case metrics.Timer:
-			ms := metric.Snapshot()
-			ps := ms.Percentiles([]float64{0.5, 0.75, 0.95, 0.99, 0.999, 0.9999})
-
-			measurement := fmt.Sprintf("%s%s.timer", namespace, name)
-			fields := map[string]interface{}{
-				"count":    ms.Count(),
-				"max":      ms.Max(),
-				"mean":     ms.Mean(),
-				"min":      ms.Min(),
-				"stddev":   ms.StdDev(),
-				"variance": ms.Variance(),
-				"p50":      ps[0],
-				"p75":      ps[1],
-				"p95":      ps[2],
-				"p99":      ps[3],
-				"p999":     ps[4],
-				"p9999":    ps[5],
-				"m1":       ms.Rate1(),
-				"m5":       ms.Rate5(),
-				"m15":      ms.Rate15(),
-				"meanrate": ms.RateMean(),
-			}
-
-			pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-			r.write.WritePoint(pt)
-
-		case metrics.ResettingTimer:
-			t := metric.Snapshot()
-
-			if len(t.Values()) > 0 {
-				ps := t.Percentiles([]float64{50, 95, 99})
-				val := t.Values()
-
-				measurement := fmt.Sprintf("%s%s.span", namespace, name)
-				fields := map[string]interface{}{
-					"count": len(val),
-					"max":   val[len(val)-1],
-					"mean":  t.Mean(),
-					"min":   val[0],
-					"p50":   ps[0],
-					"p95":   ps[1],
-					"p99":   ps[2],
-				}
-
-				pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
-				r.write.WritePoint(pt)
-			}
+		measurement, fields := readMeter(r.namespace, name, i)
+		if fields == nil {
+			return
 		}
+		pt := influxdb2.NewPoint(measurement, r.tags, fields, now)
+		r.write.WritePoint(pt)
 	})
-
 	// Force all unwritten data to be sent
 	r.write.Flush()
 }
